requirements.txt
apiclient==1.0.4
asyncio==3.4.3
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
discovery==0.0.4
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.24.0
google==3.0.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.6
oauth2client==4.1.3
oauthlib==3.2.2
pip==21.3.1
protobuf==4.25.1
pyasn1-modules==0.3.0
pyasn1==0.5.1
pyparsing==3.1.1
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests-oauthlib==1.3.1
requests==2.31.0
rsa==4.9
schedule==1.2.1
setuptools==60.2.0
six==1.16.0
soupsieve==2.5
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.37.1
